# docs/payment-orchestration.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Payment Orchestration Service API",
    "description": "Standalone payment orchestration service for AuraPoS (northflow). Manages payment merchants, provider accounts, payment intents, gateway payments, and webhooks. Phase 8K — frozen API contract.",
    "version": "0.3.0",
    "contact": {
      "name": "AuraPoS Engineering"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5100",
      "description": "Development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "serviceToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-payment-orchestration-service-token",
        "description": "Service-to-service authentication token. Required for all /v1/... routes except /v1/webhooks."
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": ["ok", "error"],
        "properties": {
          "ok": { "type": "boolean", "enum": [false] },
          "error": {
            "type": "object",
            "required": ["code", "message", "details"],
            "properties": {
              "code": { "type": "string", "example": "VALIDATION_ERROR" },
              "message": { "type": "string", "example": "name is required and must be a string" },
              "details": {
                "nullable": true,
                "description": "Structured error details (e.g. field-level validation errors). null when not applicable."
              }
            }
          }
        }
      },
...
      "Merchant": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "legalName": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["active", "suspended", "closed"] },
          "metadata": { "type": "object" }
        }
      },
      "ProviderAccount": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "merchantId": { "type": "string" },
          "provider": { "type": "string", "example": "fake_gateway" },
          "environment": { "type": "string", "enum": ["sandbox", "test", "production"] },
          "providerAccountRef": { "type": "string", "nullable": true, "description": "Provider's own account identifier. Safe to store and return in responses." },
          "status": { "type": "string", "enum": ["active", "disabled"] },
          "publicConfig": { "type": "object" },
          "metadata": { "type": "object" }
        }
      }
    },
...
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
...
      "404": {
        "description": "Not Found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
...
    }
  },
  "security": [{ "serviceToken": [] }],
  "paths": {
...
    "/v1/merchants": {
      "post": {
        "tags": ["Merchants"],
        "summary": "Create or return merchant",
        "description": "Idempotent by sourceApp + externalRef combination.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "id": { "type": "string", "description": "Optional explicit ID (UUID). Auto-generated if omitted." },
                  "name": { "type": "string" },
                  "legalName": { "type": "string", "nullable": true },
                  "sourceApp": { "type": "string", "nullable": true },
                  "externalRef": { "type": "string", "nullable": true },
                  "metadata": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Existing merchant returned", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/Merchant" } } } } } },
          "201": { "description": "New merchant created", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/Merchant" } } } } } },
          "400": { "$ref": "#/components/responses/400" }
        }
      }
    },
    "/v1/merchants/{id}": {
      "get": {
        "tags": ["Merchants"],
        "summary": "Get merchant by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Merchant found", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/Merchant" } } } } } },
          "404": { "$ref": "#/components/responses/404" }
        }
      }
    },
    "/v1/merchants/{merchantId}/provider-accounts": {
      "post": {
        "tags": ["Provider Accounts"],
        "summary": "Create provider account",
        "description": "Creates a payment provider account. `credentialsRef` accepted in body but never returned in responses.",
        "parameters": [{ "name": "merchantId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["provider", "environment"],
                "properties": {
                  "provider": { "type": "string", "example": "fake_gateway" },
                  "environment": { "type": "string", "enum": ["sandbox", "test", "production"] },
                  "providerAccountRef": { "type": "string", "nullable": true },
                  "credentialsRef": { "type": "string", "nullable": true, "description": "Env var name pointing to the secret. Never echoed in responses." },
                  "publicConfig": { "type": "object" },
                  "metadata": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Provider account created", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/ProviderAccount" } } } } } },
          "400": { "$ref": "#/components/responses/400" }
        }
      }
    },
    "/v1/merchants/{merchantId}/provider-accounts/{id}": {
      "get": {
        "tags": ["Provider Accounts"],
        "summary": "Get provider account",
        "parameters": [
          { "name": "merchantId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Provider account found", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/ProviderAccount" } } } } } },
          "404": { "$ref": "#/components/responses/404" }
        }
      }
    },
...
  }
}